Typos and format fixes in README.md
At least one typo was in a code block.

The code blocks do not have to be indented as that is just a part of the
IOCCC website and in GitHub it only makes the lines longer.

I also note that a typo is in the GitHub repo itself: in the 'about'
section we have 'oad the pwned database in a 4 level deep directory tre'
but presumably it should be either 'load' or more likely 'download' and
certainly 'tre' should be 'tree' - but if a pull request can fix that I
do not know how so I am only noting them here.